+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [0.9.31007](https://github.com/springraaar/skylobby/releases/tag/0.9.31007) - 2025-07-XX
 
-- Improved black, grey and white styles
+- Improved black, grey and light styles (javaFX renamed to Light)
 - Fix minimum size and centering for some popup windows (registration, map selection and a few others)
 
 ## [0.9.31006](https://github.com/springraaar/skylobby/releases/tag/0.9.31006) - 2025-04-22